+++ api/rs/slint/lib.rs
@@ -269,6 +269,8 @@ pub use slint_macros::slint;
 
 pub use i_slint_core::api::*;
 pub use i_slint_core::component_factory::ComponentFactory;
+#[cfg(not(target_arch = "wasm32"))]
+pub use i_slint_core::graphics::{BorrowedOpenGLTextureBuilder, BorrowedOpenGLTextureOrigin};
 pub use i_slint_core::graphics::{
     Brush, Color, Image, LoadImageError, Rgb8Pixel, Rgba8Pixel, RgbaColor, SharedPixelBuffer,
 };